# Reduce search engine memory usage