--- a/inc/commonitilobject.class.php
+++ b/inc/commonitilobject.class.php
@@ -3418,7 +3418,7 @@ abstract class CommonITILObject extends CommonDBTM {
          'datatype'           => 'itemlink',
          'searchtype'         => 'contains',
          'massiveaction'      => false,
-         'additionalfields'   => ['id', 'content', 'status']
+         'additionalfields'   => ['content', 'status']
       ];
 
       $tab[] = [