--- a/install/migrations/update_9.5.5_to_9.5.6.php
+++ b/install/migrations/update_9.5.5_to_9.5.6.php
@@ -82,6 +82,11 @@ function update955to956() {
    }
    /* /Add `date` to glpi_documents_items */
 
+   $DB->deleteOrDie('glpi_configs', [
+      'context' => 'core',
+      'name'    => 'allow_search_all',
+   ]);
+
    // ************ Keep it at the end **************
    $migration->executeMigration();