* Add LIMIT to search queries

#### Current search engine

The search engine currently only use a `LIMIT` clause if there is no search criteria specified.
In this case, it will send a second `COUNT(*)` query after the main search to know the total number of results.

Any query with a defined criteria does not use the `LIMIT` clause.
The full results are send to PHP and we only iterate on the first X results (the number of results to display on one page).

This help execution time as  `COUNT` requests can be costly but is very bad for memory consumption on big databases.

 #### Changes to the search engine

Every search request is now accompanied by a second `COUNT` request.

The mandatory clauses to re-use from the main query are `FROM`, `GROUP BY` and `HAVING`.
The `SELECT` and `FROM` clauses must only contains mandatory fields and joins to get the fastest request as possible.

The simplest case is a search without a `HAVING` clause.
This will results in a very simple `COUNT` request:
```sql
SELECT COUNT(*) FROM {main_table} {mandatory joins for the conditions fields} WHERE {condition}
```

The more complicated case is when we have a `HAVING` clause as we it forces us to add fields in the `SELECT` clause (the `HAVING` clause filters on selected results and thus cannot be done without them in the query).
In this case, we end up with a request like this:
```sql
SELECT COUNT(*) FROM (
   SELECT {having fields} FROM {main_table} {mandatory joins for the conditions fields} WHERE {condition} HAVING ... GROUP BY ...
)
```

#### Benchmark 

Here is a quick benchmark on a ~140k computers database.
We are doing some searches on computers and checking the time used by the `Search::constructData()` function as well as the memory peak of the application.
I've put in bold some values that stand out as worse.

| Search criteria | 9.5 bugfixes | #9495|
| ------------- | ---|-----|
| * | 1.18s (7.64 Mio) | 1.16s(7.60 Mio) |
| name contains 'se'  | 0.35s (**10.36 Mio**) | **0.58s** (7.62 Mio) |
| alternate username contains 'a' and location is not "Caen" | 1.51s (**56.68 Mio**)  | 1.43s (7.63 Mio) |
| manufacturer is 'HP' and status is "Status A" | 0.83s (**36.38 Mio**) | 0.73s (7.64 Mio) |
| comments contains 'ab' or serial contains '44' or uuid contains '7' | 1.78s (**72.88 Mio**)| 1.57s (7.68 Mio) |

On the memory consumption side the results are excellent.
Before these changes the memory peak will scale depending on the numbers of returned results and the number of fields displayed (more fields = more data to send back to PHP).
After these change the memory peak will always stays around the same for a given page size.

On the execution time side the results are more difficult to analyze.
The gains / loses all depend on the `COUNT` query execution time.
If it is almost instantaneous then we gain a little time or are even with the 9.5 bugfixes branches (test 1, 3, 4 and 5).
If it not instantaneous then we have worse performances, it may be a few extra tenth at best and in a few rare cases a possible x2 execution time (the `COUNT` request would be almost as long as the main query in those cases like test 2).

* Remove 'no_search' option

The search engine used this option internally to know if we had a search with no criteria and thus would be able to use the LIMIT query (as described in "Current search engine" section above).
Since we now use `LIMIT` all the time this option is no longer needed and all associated code as been removed.

* Remove 'all' criteria

As we discussed the 'all' criteria is outdated and bugged so this PR is a good opportunity to remove it.

* Fix duplicate aliases

It seems we had a few case of duplicate aliases in the `SELECT` clause of our search requests.
This is not a problem in most simple requests as MySQL allow duplicates in the results.

It is however a problem with wrapped queries like `SELECT COUNT(*) FROM ( {wrapped query} )`, duplicated alias are not allowed in the wrapped query and will cause a SQL error.

There was two kind of duplicated aliases in the code base:
One of them was on Ticket/Problem/Changes and could be fixed in the searchoption by removing an extra field.
The other was on a special case of a meta criteria join on the current itemtype (for example searching for documents on documents).
This case required a fix in the search engine by adding a "META_" prefix to meta criteria aliases in the `SELECT` and `HAVING` clauses.


| Q             | A
| ------------- | ---
| Bug fix?      | yes
| New feature?  | no
| BC breaks?    | no
| Deprecations? | no
| Tests pass?   | yes
| Fixed tickets | !22511