@AdrienClairembault 2021-09-02
Tests are failing due to a duplicate key in the COUNT query for ticket, changes and problems: 
![image](https://user-images.githubusercontent.com/42734840/131882208-b84c780d-26ee-4d43-b4a8-1b513e7e6717.png)

This is weird as the SELECT query have the same duplicate key even in 9.5/bugfixes but it doesn't trigger an error : 
![image](https://user-images.githubusercontent.com/42734840/131882426-a91fe46b-aa84-48cb-91e2-07b735befd25.png)


Probably an error in some CommonITILSearchOptions... will try to look for it tomorrow.