@AdrienClairembault 2021-09-13
Will this fix be part of 9.5.6 ?