@AdrienClairembault 2021-09-16
Regarding the latest (6984fa1) commit, it's related to another memory mismanagement that was found in the internal ref.
It turns out we are also loading **all** search results into a PHP array (so taking the results out of a memory efficient mysql resultset and putting it all in a standard PHP array row by row).

This isn't an issue when you are browsing a small sized result (e.g. a standard 20 entries result page).
However if you want to export all search results to CSV then you may reach the memory limit very easily if you have a lot of items.

This commit fix it by using a PHP generator.

The downside is that the "parsing" process of the results set can now only be used to build `$data['data']['rows']`. 
Before the changes we would also build `$data['data']['currentuser']` and `$data['data']['items']`.

-> I'm not sure what was in `$data['data']['currentuser']` but it seems to be unused data as the `'currentuser'` string isn't used anywhere else in GLPI.
-> `$data['data']['items']` seemed to contain the ids of every items returned by the searchs.
Removing it doesn't break the unit tests and I can't find any code that would depend on it so it may be unused too.
Note that the `items` key is used in a lot of place in GLPI so it's hard to check only the ones related to search results...