@AdrienClairembault 2021-09-17
The 4 latest commit reduce the memory used when exporting 140k computers to CSV from **1.4GB** to ~65MB.

65MB is still significant as there is yet another bottleneck happening simply because we load a lot of data from the database straight into PHP memory.
Getting rid of this bottleneck would require to use [unbuffered queries](https://www.php.net/manual/en/mysqlinfo.concepts.buffering.php) which would imply much bigger changes to the codebase so I don't see it happening in the near future.